Function to determine chained protocol spacer support based on iblrig version
